+++ modules/documentation/docs/changelog.md
@@ -2,8 +2,14 @@
 
 ## Next Release
 
+## 0.1.1
+
 - \[all\] Improve logging
 - \[contracts\] Add etherscan verification plugin
+- \[router\] Correctly reclaim from the sender's channel
+- \[contracts\] Check if channel is deployed right before deploying
+- \[router\] Checksum config assetIds
+- \[engine\] Checksum user input assetIds
 
 ## 0.1.0